[Kogito-8520] Separate plain builder API and Builder-k8s API
