+++ builder-k8s/build_pod.go
@@ -1,4 +1,4 @@
-package builder
+package builder_k8s
 
 import (
 	"context"
