+++ builder-k8s/builder.go
@@ -1,10 +1,10 @@
-package builder
+package builder_k8s
 
 import (
 	"context"
 	"fmt"
-
 	"github.com/kiegroup/container-builder/api"
+	//builder2 "github.com/kiegroup/container-builder/builder-k8s"
 	"github.com/kiegroup/container-builder/client"
 	"github.com/kiegroup/container-builder/util"
 	"github.com/kiegroup/container-builder/util/log"
