+++ builder-k8s/builder_test.go
@@ -1,4 +1,4 @@
-package builder
+package builder_k8s
 
 import (
 	"context"
