+++ builder-k8s/error.go
@@ -14,39 +14,32 @@
  * limitations under the License.
  */
 
-package builder
+package builder_k8s
 
 import (
 	"context"
-
 	"github.com/kiegroup/container-builder/api"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func newScheduleAction() Action {
-	return &scheduleAction{}
+func newErrorAction() Action {
+	return &errorAction{}
 }
 
-type scheduleAction struct {
+type errorAction struct {
 	baseAction
 }
 
 // Name returns a common name of the action.
-func (action *scheduleAction) Name() string {
-	return "schedule"
+func (action *errorAction) Name() string {
+	return "error"
 }
 
 // CanHandle tells whether this action can handle the build.
-func (action *scheduleAction) CanHandle(build *api.Build) bool {
-	return build.Status.Phase == api.BuildPhaseScheduling
+func (action *errorAction) CanHandle(build *api.Build) bool {
+	return build.Status.Phase == api.BuildPhaseError
 }
 
 // Handle handles the builds.
-func (action *scheduleAction) Handle(ctx context.Context, build *api.Build) (*api.Build, error) {
-	// TODO do any work required between initialization and scheduling, like enqueueing builds
-	now := metav1.Now()
-	build.Status.StartedAt = &now
-	build.Status.Phase = api.BuildPhasePending
-
-	return build, nil
+func (action *errorAction) Handle(ctx context.Context, build *api.Build) (*api.Build, error) {
+	return nil, nil
 }
