+++ builder-k8s/schedule.go
@@ -14,33 +14,38 @@
  * limitations under the License.
  */
 
-package builder
+package builder_k8s
 
 import (
 	"context"
-
 	"github.com/kiegroup/container-builder/api"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func newErrorAction() Action {
-	return &errorAction{}
+func newScheduleAction() Action {
+	return &scheduleAction{}
 }
 
-type errorAction struct {
+type scheduleAction struct {
 	baseAction
 }
 
 // Name returns a common name of the action.
-func (action *errorAction) Name() string {
-	return "error"
+func (action *scheduleAction) Name() string {
+	return "schedule"
 }
 
 // CanHandle tells whether this action can handle the build.
-func (action *errorAction) CanHandle(build *api.Build) bool {
-	return build.Status.Phase == api.BuildPhaseError
+func (action *scheduleAction) CanHandle(build *api.Build) bool {
+	return build.Status.Phase == api.BuildPhaseScheduling
 }
 
 // Handle handles the builds.
-func (action *errorAction) Handle(ctx context.Context, build *api.Build) (*api.Build, error) {
-	return nil, nil
+func (action *scheduleAction) Handle(ctx context.Context, build *api.Build) (*api.Build, error) {
+	// TODO do any work required between initialization and scheduling, like enqueueing builds
+	now := metav1.Now()
+	build.Status.StartedAt = &now
+	build.Status.Phase = api.BuildPhasePending
+
+	return build, nil
 }
